refactoring: add shared folder into components folder (it contains all components which are used everywhere in app)
