+++ src/components/shared/DeletedStamp/DeletedStamp.jsx
