+++ src/components/shared/DeletedStamp/DeletedStamp.module.scss
