+++ src/components/shared/String-button/String-button.jsx
