+++ src/screens/Colors/Person-card/PersonCard.module.scss
